+++ CHANGELOG.md
@@ -9,6 +9,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Upcoming changes...
 
+## [1.2.3] - 2023-08-25
+### Added
+- Added HEAD support for the following endpoints:
+  - `/`, `/api/`, `/api/health`, `/api/health-check`
+
 ## [1.2.2] - 2023-08-24
 ### Added
 - Added support for `/api/kb/details` endpoint
@@ -66,3 +71,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [1.2.0]: https://github.com/scanoss/api.go/compare/v1.1.0...v1.2.0
 [1.2.1]: https://github.com/scanoss/api.go/compare/v1.2.0...v1.2.1
 [1.2.2]: https://github.com/scanoss/api.go/compare/v1.2.1...v1.2.2
+[1.2.3]: https://github.com/scanoss/api.go/compare/v1.2.2...v1.2.3